improve instance handling with additional generic support
A instance can now define a keyworded property for it to act as
Instance hashs and equality have been changed to reflect this change
time builtins are making better use of this change.
concept node now better handles being called functionality
introduced decorator for better error handling
scenarios are not required for somethings!