Introduction of Count built in function
The count built in function allows a user within logic to query...
the number of instances and relations that appear in the ontology given
filters.
Efficiently filter instances for minimal evaluations
Moved Parameter to concept function into eval tree so that all nodes can use it
Moved the syntax for concept into eval tree as its fail universally required.